🎨 Network Switcher Improvements  (#2561)
* Update wallet state management.

This change introduces a new handling for wallet state that covers the cases for connecting and changing chains. This includes making use of a reducer in 'wallet-state.js' to set values for 'connecting' and 'setting-chain' states that are referenced from other components.

Additionally, the file names are updated to use kebab-case convention for file naming and improve on overall code readability for name references.

This provides a more descriptive state management for wallet connections and the ease of tracking the connections. Also, this promotes code consistency with file naming convention for the whole codebase.

* Add improved network handling in web3 onboarding

A better handling of unsupported networks was implemented in the Web3 onboarding process. It now considers wallet network related updates and takes actions accordingly. Network change rejections are now also handled properly. An unsupported network modal was introduced to alert the user when there's a network mismatch, offering them a way to switch the network. Weighting the needs of the wallet's network, some enhancements and adjustments were applied to the network switcher considering the connected network. An unsupported network state was introduced in the wallet state management.

This change makes the app's behavior more robust and user-friendly when dealing with network switches and mismatches.

* Update key property in NavigationNetworkSwitcherButton and enhance network handling in wallet state

In NavigationNetworkSwitcherButton, the key for each Link has been updated to be a combination of link.label and link.url to ensure its uniqueness.

In wallet state, when the network is disconnecting or disconnected, it now checks if the new networkHexId is supported. If it is, the state changes to 'connecting', otherwise it stays as 'unsupported-network'. This change improves handling of network transitions.

* Implement automatic adding of fork to wallet

In the web3OnBoard package, we now have the useEffect hook to automatically add the fork to the wallet when one exists. Also, the `extractUuidFromUrl` function was added in the `injected-wallet-interactions.ts` file to extract UUID from the URL, which is then used to create a more descriptive chain label. The motive behind these changes is to enhance the user experience by eliminating the need to manually add forks to the wallet.

* Refactor wallet management state using ts-pattern.

* Refactor wallet state management with better transitions

Refactored and simplified the wallet connection and chain setting state management for clear transitioning between states. The changes include eliminating unnecessary state parameters, refining action handling on 'CHANGE_CHAIN', adjusting connector handling, and updating walletStateReducer. The changes result in a streamlined state management process minimizing errors during transitions.

* Update Aave v3 strategy config loading and user proxy tracking

This revision updates the logic related to the handling of Aave v3 strategies in the system. Specifically, it revises the process of loading the strategy configuration for an Aave v3 position in the system. This is achieved by connecting the corresponding DPM proxy of the user that matches the position at hand. The user’s DPM proxies are also now tracked regularly, which ensures the ability to reflect accurate strategy details at all times.

The purpose of this is to ensure that the strategy configuration is tied to the correct user, and isn’t incorrectly associated due to changes in proxy ownership. Hence the code has been updated to accommodate this design. The overall change helps to ensure that the system maintains congruent data and behavior regarding user positions and strategies.

* Add fallback to disabled network providers

Implemented a fallback check if the requested network provider is in the disabled list. The goal is to return a provider from the disabled list if it exists. This is to enhance our support for multiple network ids and improve the robustness of the getRpcProvider function, even when the requested network id is currently disabled in the app.

* Fix typo in function name

Renamed the function 'areThePageNetoworksTheSame' to the correct name 'areThePageNetworksTheSame'.