Updated subgraphLoader to use legacy default chain (#2698)
Updated the 'useSubgraphLoader.ts' file to replace the default network id with a call to 'getLegacyDefaultChain'. This ensures that the default chain id is retrieved from the legacy default chain if no network id is provided. Also, exported 'getLegacyDefaultChain' from 'use-legacy-default-chain.tsx' for use in other parts of the application.