V3 Borrow Position Switcher (#2710)
* feat: add savePositionToDb service to aave state machine

* feat: read db type in owner page

* feat: transition between position types with full page reload after save

* chore: remove logging

* chore: add comments where appropriate

* refactor: update strategy config using Subject rather than reload

* chore: fix linting issues

* chore: address typecheck issues

* feat: respect PositionCreated event when DB unavailable

* chore: remove incorrect comment

* chore: fix getStrategyConfig test

* chore: address review points

* chore: address chainId review point - get via network arg

* chore: add earn to VaultType enum and disable header temporarily

* Refactor product type signature and usage in various files

* db migration

* Update chainId source in getOpenAaveV3PositionStateMachineServices

* Re-enable header in AaveOpenView

This commit re-enables the header in AaveOpenView which was previously disabled. The related issue https://github.com/OasisDEX/oasis-borrow/pull/2717 has been merged and resolved, hence the commented out header has been uncommented and is now in use again.

* Refactor strategyConfig implementation and availableActions in Aave module

The strategyConfig object and availableActions property have been refactored to a function with conditions enabling dynamic configuration in the Aave module. This allows for feature-flagging of actions, where additional actions can be appended depending on whether the associated feature is active. A result of this refactoring is that actions can now be dynamically determined based on product type and certain feature toggles.

---------

Co-authored-by: Jakub Swierczek <jakub.swierczek@outlook.com>