Aave history [sc-10793] (#2750)
* feat: add history card

* "Add Aave position history visualization"

This commit includes position history from Aave protocol in the application. The `AaveHistoryEvent` was added into state machine context to allow the Application to include history events in Aave module. The `PositionHistory` component is also brought into `AaveManageTabBar` container to allow displaying the history of position changes, providing users with a comprehensive view of the asset's historical transactions in the Aave protocol.

* "Add historical events tracking in Aave state machine"

This commit adds the ability to track historical events in the Aave state machine. This is achieved by adding a new entry 'PROXY_RECEIVED' and 'HISTORY_UPDATED' into the state machine and connecting it to a new service 'getAaveHistoryEvents' that promises a list of AaveHistoryEvent[]. These changes provide the functionality to track and handle historical events throughout the lifetime of a user's interaction with the Aave protocol, improving transparency and traceability for the end-user. This was necessary for enabling detailed analysis and easier debugging of the Aave protocol interactions.

* feat: add history events handler

* feat: implement get history service

* better callback handling

* feat: add labels

* feat: display proper fields

* feat: custom display for automation events

* Modified Aave history events to include additional properties

Adjusted how Aave history events are mapped to include more necessary properties in the API response. Type conversion for several properties using BigNumber was added. Additionally, a new conditional property `trigger` has been introduced to improve handling of specific event types. The types for the Position History and Aave events were updated accordingly to reflect these changes. As a result, a more comprehensive data set for Aave position history is available for further display or processing.

* Add toggle control for AaveV3History feature

Added a new feature toggle 'AaveV3History' to control display of Aave history. This toggle is checked before displaying Aave history to ensure feature availability. This provides flexibility to control the visibility of this feature and make it possible to disable or enable it without changing the code.

* max coverage

* feat: display autoamtion properties

* Expand SubgraphMethodsRecord type to span multiple lines

The type SubgraphMethodsRecord has been expanded to span multiple lines for better readability and maintenance. The previous single-line record was getting cluttered as more subgraphs were added, potentially making future additions or changes more prone to errors. This change results in a cleaner code with the ease of addition or deletion of Subgraphs.

* Lower sentry sampling rates and adjust console integration levels

Sampling rates for traces, general sampling, and profiles have been reduced to lower data noise and improve performance. Console integration now only captures 'error' level messages, for a more manageable level of verbosity. These changes aim to balance captured data volume and system performance.

* Update network context handling in PositionHistory component

Removed reliance on AppContext for network data in the PositionHistory component. NetworkId is now explicitly passed as a prop to lower coupling and improve data flow clarity. This change also simplifies its inner workings, removing the need for the useObservable hook and a Skeleton loading state. Added checks with ensureEtherscanExist and ensureEthtxExist to verify required properties are available and avoid potential runtime errors.

---------

Co-authored-by: Jakub Swierczek <jakub.swierczek@outlook.com>