Add support for optimization tab on Aave V3 strategies. (#3207)
* Add support for optimization tab on Aave V3 strategies.

This commit introduces the "isOptimizationTabEnabled" method to the Aave V3 strategy. This method checks if the optimization tab is enabled for the specific strategy type and network. This check is needed to separate the availability of the optimization tab based on the vault, as the feature is not yet supported on all vaults.

Further, it adds the feature toggles for "AUTO_BUY" and "AUTO_SELL" features under the Aave protocol in the automation feature map.

* λ two new functions: get-triggers, setup-trigger (#3263)

* Add new lambdas and local development guide

This commit introduces two new lambda functions, 'get-triggers' and 'setup-trigger', and updates the documentation to include a guide for local development. Modifications were made in 'build.mjs' and 'install.mjs' files to accommodate the new lambdas, and necessary testing configurations and package-lock.json files were added. A significant part of the change went into updating the README with new instructions for creating a new lambda and to detail instructions for local development using AWS SAM and Docker.

* Adjust Autobuy view

* Remove duplicates

* missing labels

* some changes

* add loading changes

* don't check the use MaxBuyPrice

* steps

* changes in lambdas

* changes in error handling