Update triggers configuration and improve validators (#59)
This commit adds a new 'deploy:staging' dependency in turbo.json
configuration. Also, it modifies auto-buy-validator to apply the
maxBuyPrice only when useMaxBuyPrice is true. Further, it removes the
comment from sdk import statement in summer-stack.ts, and adds default
values for minSellPrice and maxBuyPrice in the respective trigger
encoders.