Change chainId validation logic in get-migrations-function
Instead of accepting number, the chainId is now an optional string which is to be parsed into a number. This modification ensures that the customRpcUrl and chainId must be provided together and also adds a new validation to reject -1 as a chainId, thereby making sure that valid chainId and customRpcUrl are provided.