Add memoized getExecutionParams function in Aave service
The getExecutionParams function has been added to the Aave trailing stop loss service, with the purpose of more accurately calculating the dynamic execution loan-to-value (LTV) and execution price. Realignments in the 'simulatePosition' and 'validate' functions have been done to incorporate the newly minted function output. This improves the performance and precision of the Aave service.