Update logical operators in stop-loss check
The previous logical operator checking for undefined values of executionLtv and executionPrice has been modified. The operator has been changed from OR (||) to AND (&&) to ensure both variables need to be undefined before warning is logged. Also, the non-null assertion operator has been added to ensure executionPrice will not be null.