Update trigger price calculations and modify GraphQL query
Changed constant variables to let for 'maxPrice' and 'originalPrice' to handle missing or lesser 'maxPrice'. Updated code to use 'originalPrice' as 'maxPrice' in such cases. This ensures a valid 'executionPrice' and 'originalPrice' are always set in 'dynamicParams'. The 'max-price.query.graphql' file was also modified to use 'derivedPriceTimestamp_gte' and order by 'derivedPrice'.