Add context to setupTriggerFunction handlers in summerfi-api
This commit includes the addition of the 'Context' type to the import statement in index.ts of the setupTriggerFunction in the summerfi-api app. Also, it adds the context as a new parameter to the handler function and uses this context to add and append keys to the logger. This change will facilitate better troubleshooting and monitoring.