chore: [sc-14565] fix trigger removal (#95)
Story details: https://app.shortcut.com/oazo-apps/story/14565

Refactor trigger removal to only revoke allowance if last trigger

Moved `CurrentStopLoss` from types to trigger-encoders. Introduced a
conditional removal of the allowance. The modification checks if the
current trigger is the last one in the account before revoking the
allowance. This optimizes gas usage and improves performance when
multiple triggers are set for the same account.