Get Locked Collateral Function (#115)
This commit adds a new function for obtaining locked collateral across
all SummerFi positions (for Ajna, Morpho Blue, Aave & Spark). It is
added to a new stack because we need to provide this function to our
partner. Currently, it only returns data for `weETH` & `eETH`, but that
could easily change in the future.