Migrate types and functions to new shared packages
This commit rearranges the project structure by moving PositionLike types and other related types and functions that are used in multiple services to shared packages. It also updates all the relevant import paths in several files. Functions that have been moved to new packages include getAavePosition, calculateLtv, simulateAutoTakeProfit, calculateCollateralPriceInDebtBasedOnLtv and others. The aim of these migrations is to avoid code duplication and make the project structure more compact and maintainable.