Remove serverless-contracts package and update dependencies
Deleted the serverless-contracts package as it was no longer needed, and updated various imports across multiple files to reflect this change. The serverless-contracts package was duplicated in the triggers-shared library, hence it was removed to avoid redundancy and improve code maintenance. Other changes include swapping position of 'CurrentTriggerLike' import and updating powertools to version 2 in get-meta-morpho-details-function's package.json file.