ERC-4626 Positions in Portfolio (#3683)
* wip position

* erc4626 position loader

* steakhouse position label

* format fix

* fix metamorpho api urls

* Introduce fetchFromFunctionsApi and isServer helpers

This commit introduces two new helper functions: 'fetchFromFunctionsApi' and 'isServer'. The 'fetchFromFunctionsApi' function replaces a fetch call in 'getErc4626ApyParameters'. It deliberates where to send fetch requests based on the environment, helping to keep the codebase clean. Meanwhile, the 'isServer' function checks whether the code is running on a server or not, which is useful for server-side rendering purposes.

---------

Co-authored-by: Jakub Swierczek <jakub.swierczek@outlook.com>