feat: [sc-14928] [CI Task] Allow ConnectorTokens to be available through the 1inch API call (#3703)
* Update exchange lists and refactor API handler

Removed several exchanges and added others in the default liquidity providers lists. Fixed the import of 'zod' and added the handling for 'connectorTokens' when the network is mainnet in the exchange API handler. These changes aim to improve the current version of the application by maintaining up-to-date exchanges and improving the API handler logic.

* Add missing connectorTokens to query parameters

This commit updates the functionality of the exchange [[...params]].tsx file to check and add missing connectorTokens query parameters. The parameters are checked against a predefined map (connectorTokensMap) and added if found missing. This ensures all necessary parameters are available for further processing.