feat: [sc-14946] [FE] Display position Aave/Spark position from DsProxy (#3710)
* Add support for migration of external positions to Aave V3 strategy

The commit adds support for migration from an external account into Aave V3 strategy. The changes include modifications in `get-addresses.ts` and `helpers/index.ts`, mainly addition of new function `lendingProtocolToSystemKeys`. Furthermore, changes also occur in `get-migration-position-parameters.ts` to support the migration. It also adds a new file `erc20-proxy-actions.json` defining the contract formation for the ERC20 proxy actions. It updates methods in `dpm-account.ts` and `actions/aave-like/types/migrate-aave-like-parameters.ts` to support the new migration type. Finally, it modifies the `executeTransaction` method in the `blockchain/contracts` for each network to integrate the new transactions related to ERC20 proxy actions.

* Add isLocked check to Aave migration views

The logic and UI for migrating Aave has been updated to include an 'isLocked' check. This check ensures the position owner and the connected web3 account match before the user can interact with the migration components for better security. This 'isLocked' check is implemented across multiple Aave migration views.

* Improve gas estimation in Aave migration feature

Added functionality to calculate the transaction cost of the gas for Aave migration. This is achieved by using a new 'useMigrationTransactionCostWithLoading' hook, which includes fetching transaction fees and setting an error state when applicable. This provides more accurate gas estimation and enhances the user experience during the migration process.

* Update package versions and simplify code

This commit updates the versions of both "@oasisdex/addresses" and "@oasisdex/dma-library" in the package.json file. In addition, it removes unnecessary comments and improves error handling in the lendingProtocolToSystemKeys function within the "aave-like" actions. These changes aim to streamline the software's functionality and enhance code readability.

* Fix lint