Update Rays Cron
This commit introduces the 'summer-events-subgraph' package which includes configuration files, queries, schema, and index files. It also adds necessary scripts and dependencies in the package.json file. The main functionality is fetching users' point data from the SummerPointsSubgraph.
A new function 'update-rays-cron-function' has been added to the project as part of the background jobs suite. This includes the set up of relevant dependencies, scripts, linting, testing configurations and minor environment tweaks. The function currently logs a 'Hello World' message as a placeholder.