Implement points distribution and multiplier insertion in cron function
The update inserts points distribution and multipliers in the 'update-rays-cron-function'. This includes various types of multipliers like protocol boost, swap, time open, automation protection, lazy vault. If the multiplier does not exist, it gets inserted, if it does, an update operation is performed. Also, the value type in the multiplier table schema is changed from integer to decimal.