Refactor points accrual and update Rays DB setup
This commit includes changes which refactor how the points are accrued. This includes adjusting different multipliers and methodology for checking the validity of the lazy vault and automation protection. In addition, it updates the database configuration for the Rays DB setup to include the logger and create new services. It also revises the update-lock-service and adds points distribution for migrations and swaps. Some values, protocols and networks have been removed or renamed for accuracy and proper functioning of the calculation system.