Update Dockerfile and workflow file for rays deployment (#367)
This pull request includes updates to the Dockerfile and workflow file
for the rays deployment. The Dockerfile has been modified to copy the
necessary files and configure the output to standalone mode. The
workflow file has been updated to include the necessary steps for
building and pushing the Docker image. These changes are necessary for
the successful deployment of the rays application.