Add prod workflow (#383)
This pull request adds a production workflow for deploying the Rays
Dashboard. The workflow includes steps for checking out the code,
caching the turbo build setup, setting up the Node.js environment,
establishing a VPN connection, installing dependencies, prebuilding,
building, configuring AWS credentials, logging in to Amazon ECR,
building a Docker image, pushing the image to ECR, updating the ECS
service with the latest Docker image, waiting for all services to become
stable, and invalidating CloudFront. This workflow will enable the
deployment of the Rays Dashboard to the production environment.