Add RAYS_DATABASE_VIEW environment variable (#407)
This pull request adds the `RAYS_DATABASE_VIEW` environment variable to
the `.env.template` file. It also updates the `get-rays-function` and
`get-rays-leaderboard-function` to use the `RAYS_DATABASE_VIEW` variable
for selecting the appropriate database view. Additionally, the
`turbo.json` file is updated to include `RAYS_DATABASE_VIEW` in the list
of environment variables.