feat: signature + narrow types refactor (#464)
- Use a unique symbol for the interfaces/classes branding. This allows
for a more generic branding in the SDK and to standardize further the
branding
- Move the `Parameters` type from the interface files to the class files
where it belongs. The `Parameters` types need to take into account which
attributes must be passed as part of the instance construction
- Narrow down the types for the relevant interfaces (i.e.: `PoolType` in
`IPool` is just `PoolType`, but for the `ILendingPool` it becomes
`PoolType.Lending` as a type itself). Now with the classes parameters it
really cleans up the initialization of the classes hierarchy

---------

Co-authored-by: Piotr Witek <739075+piotrwitek@users.noreply.github.com>