Integrate new Armada Protocol SDK backend with the FE to get user positions and position details (#491)
## Description
Integrate new Armada Protocol SDK backend with the FE to get user
positions and position details

## Changes
- Integrate new Armada Protocol to get user positions and position
details
- Added React client library for streamlined FE integration
- Fixed FE integration with the newest SDK
- Removed maybe from getConfigurationItem and replaced with throwing an
error because it's better to find missing env in the running environment

## Benefits
- I got problems finding the missing env with exising approach in the
getConfigurationItem, thus introduced
throwing an error when reading the config is better because it
immediately shows where is the issue instead of the layers above it that
were checking for undefined value and tracking it down

## Testing
Updated automated tests
Tested manually by running the webapp 

Please review and provide any feedback or suggestions for improvement.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **New Features**
- Enhanced wallet connection and message signing capabilities in the
AccountKit component.
	- Introduced a new type for fleet configuration to improve type safety.
	- Added logging capabilities to the SDK configuration.

- **Bug Fixes**
- Improved consistency in TypeScript type definitions and interfaces for
better readability.
	- Enhanced error handling in configuration retrieval methods.

- **Chores**
- Updated project dependencies to streamline development and enhance
functionality.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->