Update SupportBox to handle external links (#718)
This pull request includes several updates to the `SupportBox` component
and its integration within the navigation of the app. The changes focus
on enhancing the user experience by updating icons, modifying link
behaviors, and conditionally displaying the support section based on the
app's version.

Enhancements to `SupportBox` component:

*
[`packages/app-earn-ui/src/components/layout/Navigation/SupportBox.tsx`](diffhunk://#diff-16e315db3499ef283d17c6dee78257d9b988253d1249949625b724aaca2598bcR7):
Updated the icon sizes and changed the icons for different support
options. Modified the link behavior to open external links in a new tab
using the new `isOutsideLink` helper function.
[[1]](diffhunk://#diff-16e315db3499ef283d17c6dee78257d9b988253d1249949625b724aaca2598bcR7)
[[2]](diffhunk://#diff-16e315db3499ef283d17c6dee78257d9b988253d1249949625b724aaca2598bcL23-R32)
[[3]](diffhunk://#diff-16e315db3499ef283d17c6dee78257d9b988253d1249949625b724aaca2598bcL53-R73)

Conditional display of support section:

*
[`packages/app-earn-ui/src/components/layout/Navigation/get-navigation-items.tsx`](diffhunk://#diff-a4b50c605462510fceb03945539d7fbadfffefae701750ff40ea17e23a0ce1acR2):
Imported `SupportBox` and conditionally included the support section in
the navigation items based on whether the app is in pre-launch version.
[[1]](diffhunk://#diff-a4b50c605462510fceb03945539d7fbadfffefae701750ff40ea17e23a0ce1acR2)
[[2]](diffhunk://#diff-a4b50c605462510fceb03945539d7fbadfffefae701750ff40ea17e23a0ce1acL61-R70)

New helper function:

*
[`packages/app-earn-ui/src/helpers/is-outside-link.ts`](diffhunk://#diff-a2ef8a7ca5f2a7f538a6686fdce297fb4d7cfd1160cca6f5f7a5f751728f9c5dR1):
Added a new helper function `isOutsideLink` to determine if a URL is
external.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **New Features**
- Enhanced support navigation with dynamic visibility based on
application launch status
	- Improved external link handling in support box
- Updated icons and URLs for support options including 1:1 call, email
support, and live chat

- **Bug Fixes**
- Corrected rendering of support navigation item based on application
state
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>