Change token selection handling in VaultManageView and ControlsDepositWithdraw components (#843)
This pull request includes several changes to improve token selection
and handling in the `VaultManageViewComponent` and related components.
The most important changes include adding a new `baseTokenOptions` to
the token selector, updating the `VaultManageViewComponent` to use these
new options, and ensuring consistency in labeling for deposit and
withdrawal actions.

Improvements to token selection:

*
[`apps/earn-protocol/components/layout/VaultManageView/VaultManageViewComponent.tsx`](diffhunk://#diff-093401e3ee1390c1d36183263365baea6f0df465c9bc2cdafe8c6205c965caccL117-R118):
Added `baseTokenOptions` to the token selector and updated the options
in the component based on the transaction type.
[[1]](diffhunk://#diff-093401e3ee1390c1d36183263365baea6f0df465c9bc2cdafe8c6205c965caccL117-R118)
[[2]](diffhunk://#diff-093401e3ee1390c1d36183263365baea6f0df465c9bc2cdafe8c6205c965caccL308-R309)
*
[`packages/app-earn-ui/src/hooks/use-token-selector.ts`](diffhunk://#diff-5b2267a597c52671eccc9a3ac1fad6f5acec6b83f14f24d7a6421c9aa42155f4L30-R39):
Introduced `baseTokenOptions` in the `useTokenSelector` hook to separate
base token options from additional token options.
[[1]](diffhunk://#diff-5b2267a597c52671eccc9a3ac1fad6f5acec6b83f14f24d7a6421c9aa42155f4L30-R39)
[[2]](diffhunk://#diff-5b2267a597c52671eccc9a3ac1fad6f5acec6b83f14f24d7a6421c9aa42155f4R96)

Consistency in labeling:

*
[`packages/app-earn-ui/src/components/organisms/ControlsDepositWithdraw/ControlsDepositWithdraw.tsx`](diffhunk://#diff-2f30b3030307c9869ebd7800a304dcbd7794100616ac90bb34281629bd144e69L86-R87):
Updated the label logic to ensure the correct label is shown for deposit
and withdrawal actions.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- **Adaptive Token Selection:** The token options now adjust dynamically
based on whether you're initiating a deposit or a withdrawal, offering a
more tailored experience.
- **Improved Labeling:** The deposit/withdraw controls have been updated
to clearly reflect the current transaction action, enhancing overall
clarity and user interaction.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->